+++ tests/sc_isolate_wallet_test.go
@@ -30,6 +30,7 @@ func TestIsolateWallet(t *testing.T) {
 	AssertGenerateKey(t, generateKeyResp1).
 		WithWalletCreation().
 		WithName(walletName).
+		WithVersion(2).
 		WithMeta(map[string]string{"name": "key-1", "role": "validation"}).
 		LocatedUnder(home)